feat(ts-renderer): add option to preserve the file structure of type definitions
Refs: #10

+++ src/renderers/ts/index.ts
@@ -1,17 +1,49 @@
-import { mkdir, writeFile } from "fs/promises"
-import { dirname } from "path"
+import Debug from "debug"
+import { mkdir, rm, writeFile } from "node:fs/promises"
+import { basename, dirname, extname, join, relative } from "node:path"
+import { fileURLToPath } from "node:url"
 import { Schema } from "../../Schema.js"
+import { groupDeclarationsBySourceUrl } from "../../schema/declarations/Declaration.js"
+import { commonPrefix } from "../../shared/utils/string.js"
 import { Output } from "../Output.js"
 import { render, TypeScriptRendererOptions } from "./render.js"
 
+const debug = Debug("tsondb:renderer:ts")
+
+const extension = ".d.ts"
+
 export const TypeScriptOutput = (options: {
   targetPath: string
-  rendererOptions?: TypeScriptRendererOptions
+  rendererOptions?: Partial<TypeScriptRendererOptions>
 }): Output => ({
   run: async (schema: Schema): Promise<void> => {
-    await mkdir(dirname(options.targetPath), { recursive: true })
-    await writeFile(options.targetPath, render(options.rendererOptions, schema.declarations), {
-      encoding: "utf-8",
-    })
+    if (options.rendererOptions?.preserveFiles === true) {
+      debug("emitting declarations to multiple files...")
+      await rm(options.targetPath, { recursive: true, force: true })
+      await mkdir(options.targetPath, { recursive: true })
+      const declarationsBySourceUrl = groupDeclarationsBySourceUrl(schema.declarations)
+      const sourceRootPath = fileURLToPath(commonPrefix(...Object.keys(declarationsBySourceUrl)))
+      debug("common source root path: %s", sourceRootPath)
+      if (sourceRootPath) {
+        for (const [sourceUrl, decls] of Object.entries(declarationsBySourceUrl)) {
+          const sourcePath = fileURLToPath(sourceUrl)
+          const relativePath = dirname(relative(sourceRootPath, sourcePath))
+          const newDir = join(options.targetPath, relativePath)
+          const newPath = join(newDir, basename(sourcePath, extname(sourcePath)) + extension)
+          await mkdir(newDir, { recursive: true })
+          await writeFile(newPath, render(options.rendererOptions, decls!), {
+            encoding: "utf-8",
+          })
+        }
+        debug("emitted declaration files to %s", options.targetPath)
+      }
+    } else {
+      debug("emitting declarations to single file...")
+      await mkdir(dirname(options.targetPath), { recursive: true })
+      await writeFile(options.targetPath, render(options.rendererOptions, schema.declarations), {
+        encoding: "utf-8",
+      })
+      debug("emitted declarations to %s", options.targetPath)
+    }
   },
 })
